+++ frontend/src/pages/organization/CertManagerOverviewPage/CertManagerOverviewPage.tsx
@@ -1,6 +1,6 @@
 import { ProjectType } from "@app/hooks/api/workspace/types";
 
-import { ProductOverviewPage } from "../SecretManagerOverviewPage/SecretManagerOverviewPage";
+import { ProductOverviewPage } from "../ProductOverviewPage";
 
 export const CertManagerOverviewPage = () => (
   <ProductOverviewPage type={ProjectType.CertificateManager} />
