+++ frontend/src/pages/organization/KmsOverviewPage/KmsOverviewPage.tsx
@@ -1,5 +1,5 @@
 import { ProjectType } from "@app/hooks/api/workspace/types";
 
-import { ProductOverviewPage } from "../SecretManagerOverviewPage/SecretManagerOverviewPage";
+import { ProductOverviewPage } from "../ProductOverviewPage";
 
 export const KmsOverviewPage = () => <ProductOverviewPage type={ProjectType.KMS} />;
