+++ frontend/src/pages/organization/SshOverviewPage/SshOverviewPage.tsx
@@ -1,5 +1,5 @@
 import { ProjectType } from "@app/hooks/api/workspace/types";
 
-import { ProductOverviewPage } from "../SecretManagerOverviewPage/SecretManagerOverviewPage";
+import { ProductOverviewPage } from "../ProductOverviewPage";
 
 export const SshOverviewPage = () => <ProductOverviewPage type={ProjectType.SSH} />;
